requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
twilio==8.11.0
openai==1.10.0
requests==2.31.0
pydantic==2.5.3
python-multipart==0.0.6
redis==5.0.1
